+++ src/ChemistryTable.cpp
@@ -35,13 +35,19 @@
 
 // Author: Lance Hepler
 
+#include "pbbam/exception/BundleChemistryMappingException.h"
 #include "ChemistryTable.h"
+#include "FileUtils.h"
+#include "pugixml/pugixml.hpp"
+#include <fstream>
+#include <map>
+#include <cstdlib>
 
 namespace PacBio {
 namespace BAM {
 namespace internal {
 
-extern const std::vector<std::array<std::string, 4>> ChemistryTable = {
+extern const ChemistryTable BuiltInChemistryTable = {
 
     // BindingKit, SequencingKit, BasecallerVersion, Chemistry
 
@@ -68,11 +74,80 @@ extern const std::vector<std::array<std::string, 4>> ChemistryTable = {
     {{"100-619-300", "100-902-100", "3.1", "S/P1-C1.2"}},
     {{"100-619-300", "100-902-100", "3.2", "S/P1-C1.2"}},
     {{"100-619-300", "100-902-100", "3.3", "S/P1-C1.2"}},
+    {{"100-619-300", "100-902-100", "4.0", "S/P1-C1.2"}},
+    {{"100-619-300", "100-902-100", "4.1", "S/P1-C1.2"}},
 
     // 3.2 ("Goat"): S/P1-C1.3
     {{"100-619-300", "100-972-200", "3.2", "S/P1-C1.3"}},
-    {{"100-619-300", "100-972-200", "3.3", "S/P1-C1.3"}}
+    {{"100-619-300", "100-972-200", "3.3", "S/P1-C1.3"}},
+    {{"100-619-300", "100-972-200", "4.0", "S/P1-C1.3"}},
+    {{"100-619-300", "100-972-200", "4.1", "S/P1-C1.3"}},
+
+    // 4.0 ("Seabiscuit"); S/P2-C2
+    {{"100-862-200", "100-861-800", "4.0", "S/P2-C2"}},
+    {{"100-862-200", "100-861-800", "4.1", "S/P2-C2"}},
+
+    // 5.0 ("Iguana"); S/P2-C2
+    {{"100-862-200", "100-861-800", "5.0", "S/P2-C2"}}
+
 };
+
+ChemistryTable ChemistryTableFromXml(const std::string& mappingXml)
+{
+    if (!FileUtils::Exists(mappingXml))
+        throw BundleChemistryMappingException(
+                mappingXml, "SMRT_CHEMISTRY_BUNDLE_DIR defined but file not found");
+
+    std::ifstream in(mappingXml);
+    pugi::xml_document doc;
+    const pugi::xml_parse_result& loadResult = doc.load(in);
+    if (loadResult.status != pugi::status_ok)
+        throw BundleChemistryMappingException(
+                mappingXml, std::string("unparseable XML, error code:") + std::to_string(loadResult.status));
+
+    // parse top-level attributes
+    pugi::xml_node rootNode = doc.document_element();
+    if (rootNode == pugi::xml_node())
+        throw BundleChemistryMappingException(mappingXml, "could not fetch XML root node");
+
+    if (std::string(rootNode.name()) != "MappingTable")
+        throw BundleChemistryMappingException(mappingXml, "MappingTable not found");
+
+    ChemistryTable table;
+    try {
+        for (const auto& childNode : rootNode) {
+            const std::string childName = childNode.name();
+            if (childName != "Mapping") continue;
+            table.emplace_back(std::array<std::string, 4>{{
+                childNode.child("BindingKit").child_value(),
+                childNode.child("SequencingKit").child_value(),
+                childNode.child("SoftwareVersion").child_value(),
+                childNode.child("SequencingChemistry").child_value()}});
+        }
+    } catch(...) {
+        throw BundleChemistryMappingException(mappingXml, "Mapping entries unparseable");
+    }
+    return table;
+}
+
+const ChemistryTable& GetChemistryTableFromEnv()
+{
+    static const ChemistryTable empty{};
+    static std::map<std::string, ChemistryTable> tableCache;
+
+    std::string chemPath;
+    const char* pth = getenv("SMRT_CHEMISTRY_BUNDLE_DIR");
+    if (pth != nullptr && pth[0] != '\0')
+        chemPath = pth;
+    else return empty;
+
+    auto it = tableCache.find(chemPath);
+    if (it != tableCache.end()) return it->second;
+
+    auto tbl = ChemistryTableFromXml(chemPath + "/chemistry.xml");
+    it = tableCache.emplace(std::move(chemPath), std::move(tbl)).first;
+    return it->second;
+}
 
 } // namespace internal
 } // namespace BAM