+++ src/FastaReader.cpp
@@ -41,12 +41,10 @@
 
 #include "pbbam/FastaReader.h"
 #include <htslib/faidx.h>
+#include <stdexcept>
 #include <fstream>
 #include <iostream>
 #include <limits>
-using namespace PacBio;
-using namespace PacBio::BAM;
-using namespace std;
 
 namespace PacBio {
 namespace BAM {
@@ -54,9 +52,9 @@ namespace internal {
 
 struct FastaReaderPrivate
 {
-    ifstream stream_;
-    string name_;
-    string bases_;
+    std::ifstream stream_;
+    std::string name_;
+    std::string bases_;
 
     FastaReaderPrivate(const std::string& fn)
         : stream_(fn)
@@ -105,23 +103,21 @@ private:
     {
         if (!stream_)
             return;
-        char c = static_cast<char>(stream_.peek());
-        string line;
-        while (c != '>') {
+        int p = stream_.peek();
+        while (static_cast<char>(p) != '>' && p != EOF) {
             if (!stream_)
                 return;
+            std::string line;
             std::getline(stream_, line, '\n');
             bases_ += line;
             if (!stream_)
                 return;
-            c = static_cast<char>(stream_.peek());
+            p = stream_.peek();
         }
     }
 };
 
 } // namespace internal
-} // namespace BAM
-} // namespace PacBio
 
 FastaReader::FastaReader(const std::string& fn)
     : d_{ new internal::FastaReaderPrivate{ fn } }
@@ -142,9 +138,9 @@ FastaReader::~FastaReader(void) { }
 bool FastaReader::GetNext(FastaSequence& record)
 { return d_->GetNext(record); }
 
-vector<FastaSequence> FastaReader::ReadAll(const string& fn)
+std::vector<FastaSequence> FastaReader::ReadAll(const std::string& fn)
 {
-    vector<FastaSequence> result;
+    std::vector<FastaSequence> result;
     result.reserve(256);
     FastaReader reader{ fn };
     FastaSequence s;
@@ -152,3 +148,6 @@ vector<FastaSequence> FastaReader::ReadAll(const string& fn)
         result.emplace_back(s);
     return result;
 }
+
+} // namespace BAM
+} // namespace PacBio