+++ src/files.cmake
@@ -16,6 +16,7 @@ set( PacBioBAM_H
     ${PacBioBAM_IncludeDir}/pbbam/BaiIndexedBamReader.h
     ${PacBioBAM_IncludeDir}/pbbam/BamReader.h
     ${PacBioBAM_IncludeDir}/pbbam/CompositeBamReader.h
+    ${PacBioBAM_IncludeDir}/pbbam/CompositeFastaReader.h
     ${PacBioBAM_IncludeDir}/pbbam/BamWriter.h
     ${PacBioBAM_IncludeDir}/pbbam/BarcodeQuery.h
     ${PacBioBAM_IncludeDir}/pbbam/Cigar.h
@@ -29,6 +30,9 @@ set( PacBioBAM_H
     ${PacBioBAM_IncludeDir}/pbbam/EntireFileQuery.h
     ${PacBioBAM_IncludeDir}/pbbam/FastaReader.h
     ${PacBioBAM_IncludeDir}/pbbam/FastaSequence.h
+    ${PacBioBAM_IncludeDir}/pbbam/FastaSequenceQuery.h
+    ${PacBioBAM_IncludeDir}/pbbam/FastqReader.h
+    ${PacBioBAM_IncludeDir}/pbbam/FastqSequence.h
     ${PacBioBAM_IncludeDir}/pbbam/FrameEncodingType.h
     ${PacBioBAM_IncludeDir}/pbbam/Frames.h
     ${PacBioBAM_IncludeDir}/pbbam/GenomicInterval.h
@@ -87,6 +91,7 @@ set( PacBioBAM_H
     ${PacBioBAM_IncludeDir}/pbbam/internal/CigarOperation.inl
     ${PacBioBAM_IncludeDir}/pbbam/internal/Compare.inl
     ${PacBioBAM_IncludeDir}/pbbam/internal/CompositeBamReader.inl
+    ${PacBioBAM_IncludeDir}/pbbam/internal/CompositeFastaReader.inl
     ${PacBioBAM_IncludeDir}/pbbam/internal/DataSet.inl
     ${PacBioBAM_IncludeDir}/pbbam/internal/DataSetBaseTypes.h
     ${PacBioBAM_IncludeDir}/pbbam/internal/DataSetBaseTypes.inl
@@ -96,6 +101,7 @@ set( PacBioBAM_H
     ${PacBioBAM_IncludeDir}/pbbam/internal/DataSetListElement.inl
     ${PacBioBAM_IncludeDir}/pbbam/internal/DataSetTypes.inl
     ${PacBioBAM_IncludeDir}/pbbam/internal/FastaSequence.inl
+    ${PacBioBAM_IncludeDir}/pbbam/internal/FastqSequence.inl
     ${PacBioBAM_IncludeDir}/pbbam/internal/Frames.inl
     ${PacBioBAM_IncludeDir}/pbbam/internal/GenomicInterval.inl
     ${PacBioBAM_IncludeDir}/pbbam/internal/Interval.inl
@@ -128,7 +134,6 @@ set( PacBioBAM_H
     ${PacBioBAM_IncludeDir}/pbbam/virtual/ZmwWhitelistVirtualReader.h
 
     # library-internal headers
-    ${PacBioBAM_SourceDir}/AssertUtils.h
     ${PacBioBAM_SourceDir}/BamRecordTags.h
     ${PacBioBAM_SourceDir}/ChemistryTable.h
     ${PacBioBAM_SourceDir}/DataSetIO.h
@@ -158,7 +163,6 @@ set( PacBioBAM_CPP
 
     ${PacBioBAM_SourceDir}/Accuracy.cpp
     ${PacBioBAM_SourceDir}/AlignmentPrinter.cpp
-    ${PacBioBAM_SourceDir}/AssertUtils.cpp
     ${PacBioBAM_SourceDir}/BaiIndexedBamReader.cpp
     ${PacBioBAM_SourceDir}/BamFile.cpp
     ${PacBioBAM_SourceDir}/BamHeader.cpp
@@ -183,6 +187,8 @@ set( PacBioBAM_CPP
     ${PacBioBAM_SourceDir}/DataSetXsd.cpp
     ${PacBioBAM_SourceDir}/EntireFileQuery.cpp
     ${PacBioBAM_SourceDir}/FastaReader.cpp
+    ${PacBioBAM_SourceDir}/FastaSequenceQuery.cpp
+    ${PacBioBAM_SourceDir}/FastqReader.cpp
     ${PacBioBAM_SourceDir}/FileProducer.cpp
     ${PacBioBAM_SourceDir}/FileUtils.cpp
     ${PacBioBAM_SourceDir}/FofnReader.cpp