+++ tests/scripts/generate_data.py
@@ -5,6 +5,7 @@ from __future__ import print_function
 import os, shutil, sys
 import StringIO
 
+# FASTA generation
 fastaSeq_1 = """TAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAAC
 AACGCAGCTCCGCCCTCGCGGTGCTCTCCGGGTCTGTGCTGAGGAGAACGCAACTCCGCCGGCGCAGGCG"""
 
@@ -13,6 +14,18 @@ AACGCAGCTCCGCCCTCGCGGTGCTCTCCGGGTCTGTGCTGAGGAGAACGCAAC"""
 
 fastaSeq_3 = """TAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAAC
 ACCCTAACCCCAACCCCAACCCCAACCCCAACCCCAACCCCAACCCTAACCCCTAACCCTAACCCT"""
+
+# FASTQ generation
+
+fastqSeq_1   = """TAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACAACGCAGCTCCGCCCTCGCGGTGCTCTCCGGGTCTGTGCTGAGGAGAACGCAACTCCGCCGGCGCAGGCG"""
+fastqQuals_1 = """[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[["""
+
+fastqSeq_2   = """TAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACAACGCAGCTCCGCCCTCGCGGTGCTCTCCGGGTCTGTGCTGAGGAGAACGCAAC"""
+fastqQuals_2 = """[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[[["""
+
+fastqSeq_3   = """TAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACCCTAACACCCTAACCCCAACCCCAACCCCAACCCCAACCCCAACCCCAACCCTAACCCCTAACCCTAACCCT"""
+fastqQuals_3 = """]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]]"""
+
 
 # file creation decorator
 def fileMaker(func):
@@ -59,7 +72,8 @@ class TestDataGenerator:
             'truncated.bam' : self.makeTruncatedBam,
             'chunking_emptyfilters.subreadset.xml'   : self.makeChunkingXml,
             'chunking_missingfilters.subreadset.xml' : self.makeChunkingXml,
-            'normal.fa' : self.makeNormalFasta
+            'normal.fa' : self.makeNormalFasta,
+            'normal.fq' : self.makeNormalFastq
         }
         self.outputSymlinks = {
             'aligned.bam'      : self.makeAlignedBamCopy,
@@ -74,6 +88,7 @@ class TestDataGenerator:
             'm150404_101626_42267_c100807920800000001823174110291514_s1_p0.2.subreads.bam.pbi' : self.makeChunkingSymlink,
             'm150404_101626_42267_c100807920800000001823174110291514_s1_p0.3.subreads.bam'     : self.makeChunkingSymlink,
             'm150404_101626_42267_c100807920800000001823174110291514_s1_p0.3.subreads.bam.pbi' : self.makeChunkingSymlink,
+            'missing_pbi.bam' : self.makeMissingPbiBam,
         }
 
     def editChunkingXml(self, outputFn, removeFiltersNode):
@@ -102,6 +117,12 @@ class TestDataGenerator:
         source = os.path.join(self.testDataDir,'chunking', outputFn)
         dest   = os.path.join(self.generatedDataDir, outputFn)
         os.symlink(source, dest)
+  
+    @fileLinker
+    def makeMissingPbiBam(self, outputFn):
+        source = os.path.join(self.testDataDir, 'phi29.bam')
+        dest   = os.path.join(self.generatedDataDir, outputFn)
+        os.symlink(source, dest)
 
     @fileMaker
     def makeChunkingXml(self, outputFn):
@@ -117,6 +138,15 @@ class TestDataGenerator:
         dest = os.path.join(self.generatedDataDir, outputFn)
         with open(outputFn, 'w') as fasta_out:
             fasta_out.write(content)
+
+    @fileMaker
+    def makeNormalFastq(self, outputFn):
+        content = ("@1\n" + fastqSeq_1 + "\n+\n" + fastqQuals_1 + "\n" +
+                   "@2\n" + fastqSeq_2 + "\n+\n" + fastqQuals_2 + "\n" +
+                   "@3\n" + fastqSeq_3 + "\n+\n" + fastqQuals_3 + "\n")
+        dest = os.path.join(self.generatedDataDir, outputFn)
+        with open(outputFn, 'w') as fastq_out:
+            fastq_out.write(content)
 
     @fileMaker
     def makeTruncatedBam(self, outputFn):